Minimal viable product
These are the files for our minimal viable product. It generates a dungeon and allows the user to move the character in the dungeon. Furthermore the game hits break whenever the player tries to go out of the dungeon. This happens because we did not program the code needed for it to be possible to enter another dungeon. The reason for this choice is that we believe that this type of implementation would be beyond what the minimal viable product requires.